refactor(hooks): type useSyncManager + relocate stats domain types
Tier 4 architecture review (PR4).

- Type the ritual instances: sync/params/FULL_SYNC_STEPS[].sync were any;
  now `type SyncInstance = ReturnType<typeof useSync>` restores the useSync
  contract at the composition root. Heterogeneous result payloads stay any[].
- Move the ~85 lines of domain interfaces out of useStats.ts into a new
  src/types/stats.ts. useStats re-exports them (back-compat), and the two
  hooks that reached into the stats-hook module just for a type
  (useLibraryCheck, useMarkAsRead → IdentifiedBooks) now import from
  ../types/stats.

No behavior change. Suite 473 green, lint clean, build OK.
