+++ backlog.md
@@ -14,7 +14,7 @@
 
 ## Stan bieżący
 
-- Wersja aplikacji: **1.67.11** (źródło prawdy: `metadata.json`; mirror w `package.json` + `package-lock.json`).
+- Wersja aplikacji: **1.67.12** (źródło prawdy: `metadata.json`; mirror w `package.json` + `package-lock.json`).
 - Branch roboczy: `claude/book-aggregator-setup-t6kfvd`. Deploy leci z `main` — zmiany
   muszą trafić na `main` (PR + merge), inaczej redeploy serwuje stary kod.
 - **Konwencja PR/issue**: jedna logiczna zmiana = jeden granularny PR (nie batchujemy).
@@ -69,6 +69,13 @@
 
 Wersja ze źródła prawdy `metadata.json` (mirror w `package.json`). Najnowsze na górze.
 
+- **1.67.12** — **[Tier 4, PR4] Typowanie `useSyncManager` + relokacja typów domenowych stats.** (1) `sync: any`/
+  `FULL_SYNC_STEPS[].sync: any`/`clearOthers(any)`/`handleSyncAction(any)` → `type SyncInstance =
+  ReturnType<typeof useSync>` (kontrakt rytuału już nie wymazany). Payloady wyników zostają `any[]` (heterogen).
+  (2) ~85 linii interfejsów domenowych przeniesione z `useStats.ts` do nowego `src/types/stats.ts`; `useStats`
+  re-eksportuje je (back-compat), a dwa hooki sięgające po TYP do sąsiedniego hooka (`useLibraryCheck`,
+  `useMarkAsRead` — `IdentifiedBooks`) importują teraz z `../types/stats`. Zero zmiany zachowania. Suite 473
+  zielone, lint czysty, build OK.
 - **1.67.11** — **[Tier 4, PR3] Wspólne prymitywy request w hookach.** `src/utils/http.ts`: `markReadRequest`
   (jeden transport mark/unmark — dotąd 3 kopie: `useMarkRead`, `useMarkAsRead`, `useCyclesHarvest`; każdy hook
   nakłada swój optimistic/lock/error) + `postStop` (fire-and-forget POST `/stop` — dotąd 3 kopie w
