+++ package.json
@@ -1,7 +1,7 @@
 {
   "name": "react-example",
   "private": true,
-  "version": "1.67.11",
+  "version": "1.67.12",
   "type": "module",
   "engines": {
     "node": ">=18 <21"
